api/requirements.txt
asgiref==3.2.10
Django==3.0.8
gunicorn==20.0.4
djangorestframework==3.11.0
Pygments==2.6.1
pytz==2020.1
sqlparse==0.3.1